Update style.css
Created variable colour for top nav buttons. Added padding and changed hover colours to categorised different media blocks. Applied css to top nav buttons and aligned them to the center. Applied a hover animation to the blocks so as to align and tie in with the animation in the header.